Split Checks into separate files
